Win32 transparent window: Fix click through
Before: When a transparent window is in the foreground
and you try to click anywhere in the transparent area,
the transparent window will register the click instead
of the underneath window.

Now: A transparent window is now click through.

Note 0: Maybe another hint should be made to allow non-
click-through transparent windows.

Note 1: Clicking the visible area of a transparent window
will also be click through. This shouldn't be an issue
for transparent windows used as overlays or fancy splash screens.